Remove Singleton design pattern. Add error message when submiting only "Check" in console for invalid command. Add error message when writing only the name of the check command.
Remove Singleton design pattern from project because it is not possible to implement it properly at the moment. Submitting "Check" in the console was throwing a MissingMethodException because it was trying to instantiate an abstract class. Now the exception is catched and an invalid command error message is printed in the console.